actualizacion

+++ src/DAO/Controllers/controller.products.js
@@ -3,12 +3,12 @@ const Products = require("../../models/Products.model");
 const Cart = require("../../models/Carts.model");
 const router = Router();
 const privateAccess = require("../../middlewares/privateAccess.middleware");
-const productSearch = require("../products.dao");
+const productsDao = require("../products.dao");
 const adminAccess = require("../../middlewares/adminAccess.middleware.js");
 const userAcces = require("../../middlewares/userAccess.middleware.js");
-const ProductsRepository = require("../repository/products.repository");
 const logger = require("../../config/logs/logger.config");
-const ErrorRepository = require("../repository/errors.repository");
+const ErrorRepository = require("../repository/error.repository");
+const mailerDao = require("../mailer.dao");
 
 router.get("/", privateAccess, async (req, res, next) => {
   try {
@@ -16,9 +16,11 @@ router.get("/", privateAccess, async (req, res, next) => {
     const message = user
       ? `Bienvenido ${user.role} ${user.first_name} ${user.last_name}!`
       : null;
+
     const cart = await Cart.findOne({ _id: user.cartId });
+
     const cartId = cart._id.toString();
-    const products = await productSearch(req, message, cartId);
+    const products = await productsDao.searchProducts(req, message, cartId);
 
     logger.info("Productos cargados con exito", products);
 
@@ -31,8 +33,7 @@ router.get("/", privateAccess, async (req, res, next) => {
 
 router.get("/mockingProducts", userAcces, async (req, res, next) => {
   try {
-    const productsRepository = new ProductsRepository();
-    const mockProducts = await productsRepository.generateMockProducts();
+    const mockProducts = await productsDao.generateMockProducts();
     res.json({ Productos: mockProducts });
   } catch (error) {
     logger.error("Error al generar los productos", error);
@@ -72,12 +73,10 @@ router.put("/:productId", adminAccess, async (req, res, next) => {
       { new: true }
     );
     logger.info("Producto actualizado con exito", updatedProduct);
-    res
-      .status(200)
-      .json({
-        message: "Product updated successfully",
-        product: updatedProduct,
-      });
+    res.status(200).json({
+      message: "Product updated successfully",
+      product: updatedProduct,
+    });
   } catch (error) {
     logger.error("Error al actualizar el producto");
     next(error);
@@ -86,11 +85,32 @@ router.put("/:productId", adminAccess, async (req, res, next) => {
 
 router.delete("/:productId", adminAccess, async (req, res, next) => {
   try {
-    await Products.findByIdAndDelete(req.params.productId);
-    logger.info("Producto eliminado", req.params.productId);
-    res.json({
-      message: `Product with ID ${req.params.productId} has been deleted`,
-    });
+    const product = await Products.findById(req.params.productId);
+    const user = req.session.user;
+
+    if (user.role === "premium") {
+      const mailOptions = {
+        from: "diegoedvflores03@gmail.com",
+        to: user.email,
+        subject: "Producto Eliminado",
+        text: `Tu producto ${product.name} fue eliminado por terminos de privacidad`,
+      };
+
+      await mailerDao.sendMail(mailOptions);
+    }
+
+    if (user.role === "administrador" || user.role === product.owner) {
+      await Products.findByIdAndDelete(req.params.productId);
+      logger.info("Producto eliminado", req.params.productId);
+      res.json({
+        message: `Product with ID ${req.params.productId} has been deleted`,
+      });
+    } else {
+      throw new ErrorRepository(
+        "No tienes permiso para eliminar este producto",
+        401
+      );
+    }
   } catch (error) {
     logger.error("Error al eliminar el producto", error);
     next(error);
